Made setAmount take in proper type

+++ android_studio_project/app/src/main/java/com/git_er_done/cmput301f22t06_team_project/models/Ingredient.java
@@ -94,7 +94,7 @@ public abstract class Ingredient {
         return amount;
     }
 
-    public void setAmount(Integer amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
